test(cli): split code-agent-loop + worker-run test monoliths under 900 lines
The cli source was de-monolithed (code.ts 884, worker-run.ts 848) but the two
biggest hand-written files in the whole repo were their TESTS — code-agent-loop
.test.ts (1848) and worker-run.test.ts (1260), which grew with the v0.80.x CLI
Fleet work. Split them by concern so no new test monolith ships in v0.80.3:

- code-agent-loop.test.ts (562) — core loop & approvals
- code-agent-loop-tools.test.ts (509) — tool calls & repair
- code-agent-loop-media.test.ts (625) — multimodal & BYOK subprocess
- code-agent-loop-resume.test.ts (473) — resume & checkpoint
- worker-run.test.ts (583) — brief parsing & external adapters
- worker-run-fleet.test.ts (710) — run, diff & gate

Each new file keeps the original imports/helpers (self-contained) and a
`code-agent-loop`/`worker-run` filename prefix so `test:cli-fleet`'s name filter
still catches all of them. Pure test reorg — no source changed; the 66 tests are
preserved exactly (every it() lands in one slice). cli typecheck 0, 66 pass.
